basis.type: renamed enumeration to enum
