+++ src/basis/type.js
@@ -76,7 +76,7 @@ module.exports = {
   string: require('./type/string.js'),
   number: require('./type/number.js'),
   int: require('./type/int.js'),
-  enumeration: require('./type/enumeration.js'),
+  'enum': require('./type/enum.js'),
   array: require('./type/array.js'),
   object: require('./type/object.js'),
   date: require('./type/date.js'),
