+++ src/basis/type/enum.js
